useSelectedStakeholder hook
